Moved ClimateImpact into App
