+++ apps/site-app/src/containers/Widgets/ClimateImpact/index.tsx
@@ -1,12 +1,12 @@
 import React from "react";
-import { ClimateImpact } from "@thecointech/shared/containers/Widgets/ClimateImpact";
+import { ClimateImpact } from "./Widget";
 import styles from './styles.module.less';
 
 export const ClimateImpactWidget = () => {
 
   return (
     <div className={`${styles.columnRightBottom} x2spaceBefore`}>
-        <ClimateImpact/>
+      <ClimateImpact />
     </div>
   )
 }
